Reuse some code between video and image.
