+++ assets/src/edit-story/elements/image/display.js
@@ -27,10 +27,10 @@ import { useRef } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import { elementFillContent } from '../shared';
+import { elementFillContent, getMediaProps } from '../shared';
 import { useTransformHandler } from '../../components/canvas';
 import StoryPropTypes from '../../types';
-import { imageWithScale, getImgProps, getImageWithScaleCss } from './util';
+import { imageWithScale, getImageWithScaleCss } from './util';
 
 const Element = styled.div`
 	${ elementFillContent }
@@ -49,7 +49,7 @@ function ImageDisplay( {
 	const imageRef = useRef( null );
 
 	// eslint-disable-next-line @wordpress/no-unused-vars-before-return
-	const imgProps = getImgProps( width, height, scale, focalX, focalY, origRatio );
+	const imgProps = getMediaProps( width, height, scale, focalX, focalY, origRatio );
 
 	useTransformHandler( id, ( transform ) => {
 		const target = imageRef.current;
@@ -58,7 +58,7 @@ function ImageDisplay( {
 		} else {
 			const { resize } = transform;
 			if ( resize[ 0 ] !== 0 && resize[ 1 ] !== 0 ) {
-				const newImgProps = getImgProps( resize[ 0 ], resize[ 1 ], scale, focalX, focalY, origRatio );
+				const newImgProps = getMediaProps( resize[ 0 ], resize[ 1 ], scale, focalX, focalY, origRatio );
 				target.style.cssText = getImageWithScaleCss( newImgProps );
 			}
 		}
